Fixes, improvements, new scripts
## Changes
Standardised console warning and info messages to be prefixed with the script filename using the symbol `__FILE__`
## Fixes
### teleport.nut
If the target is an instance, it's checked to see if it's an instance of CBaseEntity or Players (my list thing). Previously it was doing IsValid() on Players, causing an error.
## New
### barrels.nut
WIP script for Castle Steveh 3 which spawns items when a barrel is broken. Items so far include:
* A bomb which pops upward and outward from the ground
* Some money!
The script also enables the money HUD element and saves spawn count to file. The former functionality will probably be moved elsewhere in future and the latter functionality was just for experimentation.
### bumpercars.nut
WIP script which handles everything to do with starting and stopping a bumper car race. Currently working on checkpoint triggers.