New and updated scripts
### New: melee_only 0.1.1
Put players into and out of melee-only. Accounts for situations where a weapon would ordinarily prevent or complicate switching to melee, e.g. buff banner.
Useable when added to a trigger, or via RunScriptCode inputs to a logic_script.
### runspeed 0.2
Set player run speed.
Changed how it works to make it easier to use.
Functions are now methods of CTFPlayer. This means you can call them on players directly, either in response to activating an I/O chain or as inputs to player entities.
Set speed in units per second or as a decimal of a class's run speed. e.g. 0.5 is 50%, to simulate a stun with no effects.
### New: hide_items 0.4
Hide or kill a player's weapons or wearables. Useful when giving them a custom model.
### New: personal_music
Play a sound or music track to a player so only they can hear it. Play another track and it will stop the existing one and play the new one. Play the same track again and the request will be ignored, enabling easier trigger logic for environmental music. Tracks are stopped on round restart. Please read the note on caching long pieces of music.
### thirdperson 2.3
Easily put players into and out of thirdperson. Useful in platforming minigames.
Like two scripts previously, thirdperson functions are now added to CTFPlayer as methods. This lets you put all players into thirdperson easily using a single output rather than having to rely on them touching a trigger, which is less convenient.